refactor(client-presence-runtime): move SystemWorkspaceDatastore to protocol.ts - clean up tests to use self-references
